fix some general things

+++ bonus/src/raycasting/textures_bonus.c
@@ -6,7 +6,7 @@
 /*   By: mberila <mberila@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/25 15:34:57 by mberila           #+#    #+#             */
-/*   Updated: 2025/10/10 15:35:52 by mberila          ###   ########.fr       */
+/*   Updated: 2025/10/15 15:18:24 by mberila          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -72,7 +72,7 @@ void	render_textured_column(t_ray *ray, int screen_x, double line_h)
 		data()->wall_yp = data()->dist_from_top / line_h;
 		data()->ty = (int)(data()->wall_yp * (double)data()->tex->height);
 		check_boundaries();
-		data()->color = (int)sample_rgba(data()->tex, data()->tx, data()->ty);
+		data()->color = sample_rgba(data()->tex, data()->tx, data()->ty);
 		if ((data()->color & 0xFF) != 0)
 			draw_pixel(screen_x, data()->y, data()->color);
 	}
